moved pages to src
